+++ include/contipede_platform.h
@@ -5,7 +5,7 @@
 void cont_plat_waitms(int numMS); // stops execution for a certain number of miliseconds
 
 int cont_plat_get_frametime();
-int cont_plat_speed_increment(int input);
-int cont_plat_timeout_from_speed(int speed);
+int cont_plat_speed_increment(double input);
+int cont_plat_timeout_from_speed(double speed);
 
 void cont_plat_beep(); // must be done inside "draw"
